backend/requirements.txt
fastapi[standard]==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
numpy==1.26.4
scikit-learn==1.4.2
river==0.22.0
python-dotenv==1.0.1

torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
open-clip-torch==2.24.0
pillow==10.3.0
